+++ animeList.js
@@ -1,0 +1,9 @@
+module.exports = { list: function(){
+    
+    const list = '1 Death Note. 9.98 / 10\n' + 
+    '2 Fullmetal Alchemist: Brotherhood. 9.59 / 10\n' +
+    '3 Naruto. 9.31 / 10\n' +
+    '4 Attack on Titan. 9.74 / 10.\n' +
+    '5 Dragon Ball Z. 9.15 / 10';
+    return list;
+}};
