+++ index.js
@@ -1,9 +1,13 @@
 require('dotenv').config();
+var inf = require('./info');
+var dani = require('./dani');
+var cringe = require('./animeList');
+var playlist = require('./chill');
 const Discord = require('discord.js');
 const client = new Discord.Client();
 const prefix = '-';
 
-const lastBan = new Date('2/5/2021');
+
 client.once('ready', () => {
     console.log('DaniBot is online!');
 });
@@ -15,39 +19,22 @@ client.on('message', message => {
     const command = args.shift().toLowerCase();
 
     if (command === 'dani') {
-        const today = new Date();
-        const diffTime = Math.abs(lastBan - today);
-        const diffDays = Math.ceil((diffTime + (1000*60*60*3)) / (1000 * 60 * 60 * 24));
-        console.log(today);
-        message.channel.send(diffDays-1 + ' days since the last ban! Keep it up 🙂');
+        message.channel.send(dani.lastBan());
     }
     else if (command == 'tilted') {
         message.channel.send('https://on-winning.com/avoid-tilt/');
     }
     else if (command == 'cringe') {
-        const list = '1 Death Note. 9.98 / 10\n' + 
-                     '2 Fullmetal Alchemist: Brotherhood. 9.59 / 10\n' +
-                     '3 Naruto. 9.31 / 10\n' +
-                     '4 Attack on Titan. 9.74 / 10.\n' +
-                     '5 Dragon Ball Z. 9.15 / 10';
-        message.channel.send(list);
+      
+        message.channel.send(cringe.list());
     }
     else if (command=='info'){
-        const commands = 'Supported Commands:\n'+
-                        '-Dani - Days since last ban\n' + 
-                        '-tilted - Article for avoiding tilt \n' + 
-                        '-cringe - C R I N G E (Top 5 An*mes)\n' +
-                        '-chill - TOP VIBES\n' +
-                        '-stream @username - PUSKAI STREAM WE\n' +
-                        '-rank1 vlad - RANK1 VLAD HEAVY 🧢 \n' +
-                        '-mm - Check if you are in Losers or Winners Queue(using Algebra 2 and algorithms from NASA)';
-
-        message.channel.send(commands);
+        message.channel.send(inf.info());
     }
     else if (command == 'chill')
     {   
-        const command = 'https://www.youtube.com/watch?v=einR_rqDICU&ab_channel=YordanAngelov';
-        message.channel.send(command);
+        
+        message.channel.send(playlist.chill());
     }
     else if (command == 'stream'){
         const user = args[0];
@@ -63,6 +50,13 @@ client.on('message', message => {
         var number = (Math.floor(Math.random() * 100))%2;
         var options = ["Losers Queue","Winners Queue"];
         message.channel.send(options[number]);
+    }
+    else if(command == 'newyear'){
+        const command = 'https://pubmed.ncbi.nlm.nih.gov/7396691/'
+        
+        message.channel.send('Не сте сами! <@!214072494737457152>  <@!374199399146061836> ');
+       
+        message.channel.send(command);
     }
 });
 client.login(process.env.token);
