+++ info.js
@@ -1,0 +1,14 @@
+
+module.exports = { info: function(){
+
+    const commands = 'Supported Commands:\n'+
+    '-Dani - Days since last ban\n' + 
+    '-tilted - Article for avoiding tilt \n' + 
+    '-cringe - C R I N G E (Top 5 An*mes)\n' +
+    '-chill - TOP VIBES\n' +
+    '-stream @username - PUSKAI STREAM WE\n' +
+    '-rank1 vlad - RANK1 VLAD HEAVY 🧢 \n' +
+    '-mm - Check if you are in Losers or Winners Queue(using Algebra 2 and algorithms from NASA)';
+
+    return commands;
+}};
